app/requirements.txt
python-dotenv
fastapi
faster-whisper
openai
pydantic
python-multipart
uvicorn